A lot of work in adding service layer to seperate controller from repositories.
