+++ src/main/java/com/stanostrovskii/service/AdminRoomService.java
@@ -1,0 +1,8 @@
+package com.stanostrovskii.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AdminRoomService {
+	
+}
